# [Coro] Use DebugInfoCache to speed up cloning in CoroSplitPass